% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_df.R
\docType{data}
\name{nps_df}
\alias{nps_df}
\title{Example data}
\format{A data frame with 10 rows and 11 variables:
\describe{
  \item{unit}{category}
  \item{satisfaction_0_10}{Satisfaction, scale 0 - 10}
}}
\usage{
nps_df
}
\description{
A dataframe that contains dummy data in wide format for computing Attitudinal Equity
}
\keyword{datasets}
